finish all sup working and prepared 19k ul data for ssl. Need worknig on kd in a new file.
